+++ package.json
@@ -37,6 +37,8 @@
     "@types/mongoose": "^5.7.1",
     "@nestjs/typeorm": "^6.1.2",
     "@types/node": "^13.1.6",
+    "@nestjs/jwt": "^6.1.1",
+    "@nestjs/passport": "^6.1.0",
     "@types/supertest": "^2.0.8",
     "@typescript-eslint/eslint-plugin": "^2.12.0",
     "@typescript-eslint/parser": "^2.12.0",
@@ -49,12 +51,16 @@
     "luxon": "^1.17.2",
     "jest": "^24.9.0",
     "prettier": "^1.18.2",
+    "passport": "^0.4.0",
+    "passport-jwt": "^4.0.0",
     "supertest": "^4.0.2",
     "ts-jest": "25.1.0",
     "ts-loader": "^6.2.1",
     "ts-node": "^8.6.0",
     "tsconfig-paths": "^3.9.0",
-    "typescript": "^3.7.4"
+    "typescript": "^3.7.4",
+    "sha1": "^1.1.1",
+    "yup": "^0.27.0"
   },
   "jest": {
     "moduleFileExtensions": [
